Added New Task functionality on front-end
Created New Task form using react bootstrap form and useState hooks. The multi-select drop down was created using a library called react-select, as it was very difficult to create one from scratch. This form may be re-used for the edit task functionality. This currently only runs off test data and printing the inputs to a console