Added new page for editing tasks and functionality to delete a task
The delete task function is no available via a tasks large view on the dashboard page, this will also cause the data to be refreshed to show the change. The edit task page function is also available via a tasks large view which will go to a new page when edits can be made and will send the user back to the dashboard once complete.