Added editTask data check and moved to useContext for login state
editTask page now checks if there is data, in case a user tries to access the url by typing it in the user will be sent back to the dashboard. Instead of using local host to hold the state of if the user is logged in, the app.jsx uses a  context which provides the state of the user and a function to login and out. This is passed to all routes and nav bar layout thus acts similar to a global variable which is much easier to keep track of, the app.jsx calls the authentication check for the cookies rather than having to call the checks from every page. Now the nav bar can access this state the next step will be changing the available links based on if the user is logged in or not